requirements.txt
streamlit
numpy
pandas
aiohttp
asyncio
openai
python-dotenv
